+++ src/libbiguint/bu_scan.c
@@ -5,14 +5,15 @@
 #define _BSD_SOURCE
 #endif
 
+#include <sys/types.h>
+#include <stdint.h>
 #include <errno.h>
-#include <skalibs/uint32.h>
 #include <skalibs/bitarray.h>
 #include <skalibs/biguint.h>
 
-int bu_scan (char const *s, unsigned int len, uint32 *x, unsigned int xn, unsigned int zeron)
+int bu_scan (char const *s, size_t len, uint32_t *x, unsigned int xn, size_t zeron)
 {
-  register unsigned int n = bitarray_div8(zeron) ;
+  register size_t n = bitarray_div8(zeron) ;
   if (xn < n) return (errno = EOVERFLOW, 0) ;
   bu_scan_internal(s, len, x) ;
   bu_zero(x + n, xn - n) ;
